refactor: move `commands` to `discord.commands`
