+++ src/main/java/pw/mihou/amelia/db/methods/feeds/Feeds.kt
@@ -5,9 +5,9 @@ import org.javacord.api.entity.channel.ServerTextChannel
 import org.javacord.api.entity.server.Server
 import org.javacord.api.entity.user.User
 import org.javacord.api.interaction.SlashCommandInteractionOption
-import pw.mihou.amelia.commands.templates.TemplateMessages
 import pw.mihou.amelia.db.FeedDatabase
 import pw.mihou.amelia.db.models.FeedModel
+import pw.mihou.amelia.discord.commands.templates.TemplateMessages
 import pw.mihou.amelia.rss.reader.RssReader
 import pw.mihou.models.user.UserResultOrAuthor
