+++ src/main/java/pw/mihou/amelia/discord/DiscordClient.kt
@@ -9,15 +9,15 @@ import org.javacord.api.entity.intent.Intent
 import org.javacord.api.listener.connection.ReconnectListener
 import org.javacord.api.listener.connection.ResumeListener
 import org.javacord.api.listener.server.ServerLeaveListener
-import pw.mihou.amelia.commands.AnnounceCommand
-import pw.mihou.amelia.commands.FeedSubscriptionCommand
-import pw.mihou.amelia.commands.FeedsCommand
-import pw.mihou.amelia.commands.PingCommand
-import pw.mihou.amelia.commands.RegisterCommand
-import pw.mihou.amelia.commands.RemoveCommand
-import pw.mihou.amelia.commands.TestCommand
 import pw.mihou.amelia.configuration.Configuration
 import pw.mihou.amelia.db.MongoDB
+import pw.mihou.amelia.discord.commands.AnnounceCommand
+import pw.mihou.amelia.discord.commands.FeedSubscriptionCommand
+import pw.mihou.amelia.discord.commands.FeedsCommand
+import pw.mihou.amelia.discord.commands.PingCommand
+import pw.mihou.amelia.discord.commands.RegisterCommand
+import pw.mihou.amelia.discord.commands.RemoveCommand
+import pw.mihou.amelia.discord.commands.TestCommand
 import pw.mihou.amelia.discord.delegates.NexusConfigurator
 import pw.mihou.amelia.discord.delegates.ReaktConfigurator
 import pw.mihou.amelia.listeners.AnnouncementModalListener
