+++ src/main/java/pw/mihou/amelia/discord/commands/RegisterCommand.kt
@@ -1,13 +1,13 @@
-package pw.mihou.amelia.commands
+package pw.mihou.amelia.discord.commands
 
 import org.javacord.api.entity.channel.ChannelType
 import org.javacord.api.interaction.SlashCommandOption
 import org.javacord.api.interaction.SlashCommandOptionType
-import pw.mihou.amelia.commands.middlewares.Middlewares
-import pw.mihou.amelia.commands.subcommands.RegisterAuthorSubcommand
-import pw.mihou.amelia.commands.subcommands.RegisterHelpSubcommand
-import pw.mihou.amelia.commands.subcommands.RegisterListSubcommand
-import pw.mihou.amelia.commands.templates.TemplateMessages
+import pw.mihou.amelia.discord.commands.middlewares.Middlewares
+import pw.mihou.amelia.discord.commands.subcommands.RegisterAuthorSubcommand
+import pw.mihou.amelia.discord.commands.subcommands.RegisterHelpSubcommand
+import pw.mihou.amelia.discord.commands.subcommands.RegisterListSubcommand
+import pw.mihou.amelia.discord.commands.templates.TemplateMessages
 import pw.mihou.nexus.features.command.facade.NexusCommandEvent
 import pw.mihou.nexus.features.command.facade.NexusHandler
