+++ src/main/java/pw/mihou/amelia/discord/commands/RemoveCommand.kt
@@ -1,14 +1,14 @@
-package pw.mihou.amelia.commands
+package pw.mihou.amelia.discord.commands
 
 import java.awt.Color
 import org.javacord.api.entity.message.embed.EmbedBuilder
 import org.javacord.api.interaction.SlashCommandOption
 import org.javacord.api.interaction.SlashCommandOptionType
-import pw.mihou.amelia.commands.middlewares.Middlewares
-import pw.mihou.amelia.commands.templates.TemplateMessages
 import pw.mihou.amelia.db.FeedDatabase
 import pw.mihou.amelia.db.methods.feeds.Feeds
 import pw.mihou.amelia.db.models.FeedModel
+import pw.mihou.amelia.discord.commands.middlewares.Middlewares
+import pw.mihou.amelia.discord.commands.templates.TemplateMessages
 import pw.mihou.amelia.utility.confirmationMenu
 import pw.mihou.amelia.utility.redactListLink
 import pw.mihou.nexus.features.command.facade.NexusCommandEvent
