+++ src/main/java/pw/mihou/amelia/discord/commands/components/ErrorEmbed.kt
@@ -1,19 +1,20 @@
 @file:Suppress("NAME_SHADOWING")
 
-package pw.mihou.amelia.commands.components
+package pw.mihou.amelia.discord.commands.components
 
 import pw.mihou.reakt.Reakt
 import pw.mihou.reakt.elements.Embed
 
 @Suppress("ktlint:standard:function-naming")
-fun Reakt.Document.PlainEmbed(msg: String) =
-    component("pw.mihou.amelia.PlainEmbed") {
+fun Reakt.Document.ErrorEmbed(msg: String) =
+    component("pw.mihou.amelia.ErrorEmbed") {
         val msg: String by ensureProp()
         render {
             Embed {
                 Body {
                     msg.append
                 }
+                Color(java.awt.Color.RED)
             }
         }
     }("msg" to msg)
