+++ src/main/java/pw/mihou/amelia/discord/commands/components/PaginationButtonsComponent.kt
@@ -1,6 +1,6 @@
 @file:Suppress("NAME_SHADOWING")
 
-package pw.mihou.amelia.commands.components
+package pw.mihou.amelia.discord.commands.components
 
 import org.javacord.api.entity.user.User
 import pw.mihou.reakt.Reakt
