+++ src/main/java/pw/mihou/amelia/discord/commands/components/PlainEmbed.kt
@@ -1,20 +1,19 @@
 @file:Suppress("NAME_SHADOWING")
 
-package pw.mihou.amelia.commands.components
+package pw.mihou.amelia.discord.commands.components
 
 import pw.mihou.reakt.Reakt
 import pw.mihou.reakt.elements.Embed
 
 @Suppress("ktlint:standard:function-naming")
-fun Reakt.Document.ErrorEmbed(msg: String) =
-    component("pw.mihou.amelia.ErrorEmbed") {
+fun Reakt.Document.PlainEmbed(msg: String) =
+    component("pw.mihou.amelia.PlainEmbed") {
         val msg: String by ensureProp()
         render {
             Embed {
                 Body {
                     msg.append
                 }
-                Color(java.awt.Color.RED)
             }
         }
     }("msg" to msg)
